Add checks for required options
Yes, that sounds a bit odd, but sf has things like the application and
the connection as options even though they are required and would better
fit as parameters.
Anyway, in order to avoid invalid calls to the task, it now checks to
make sure everything it needs is set.